separate widget and add test for it
