+++ src/lib/installation/dialogs/security.rb
@@ -20,6 +20,7 @@
 require "yast"
 require "cwm/dialog"
 require "installation/widgets/selinux_policy"
+require "installation/widgets/polkit_default_priv"
 
 Yast.import "Hostname"
 Yast.import "Mode"
@@ -143,53 +144,6 @@ module Installation
             )
           )
         )
-      end
-    end
-
-    class PolkitDefaultPriv < CWM::ComboBox
-      def initialize(settings)
-        textdomain "installation"
-
-        @settings = settings
-      end
-
-      def label
-        _("PolicyKit Default Priviledges")
-      end
-
-      def items
-        @settings.human_polkit_priviledges.to_a
-      end
-
-      def help
-        _(
-          "<p>SUSE ships with three sets of default privilege " \
-          "settings. These are as follows:<br><ul>" \
-          "<li>\"restrictive\": conservative settings that " \
-          "require the root user password for a lot of actions" \
-          " and disable certain actions completely for remote " \
-          "users.</li>" \
-          "<li>\"standard\": balanced settings that restrict " \
-          "sensitive actions to require root authentication " \
-          "but allow less dangerous operations for regular " \
-          "logged in users.</li>" \
-          "<li>\"easy\": settings that are focused on ease " \
-          "of use. This sacrifices security to some degree " \
-          "to allow a more seamless user experience without" \
-          " interruptions in the workflow due to password " \
-          "prompts.</li></ul><br>" \
-          "The \"default\" is to keep value empty and it will be" \
-          "assigned automatically.</p>"
-        )
-      end
-
-      def init
-        self.value = @settings.polkit_default_proviledges || "default"
-      end
-
-      def store
-        res = value == "default" ? nil : value
-        @settings.polkit_default_proviledges = res
       end
     end
   end
